+++ tests/unit/test_charm.py
@@ -26,7 +26,6 @@ SAMPLE_CONFIG: JujuConfig = {
     state.PATH_CONFIG_NAME: "/",
     state.BACKENDS_CONFIG_NAME: "10.10.1.1,10.1.1.2",
     state.PROTOCOL_CONFIG_NAME: "https",
-    state.HEALTH_CHECK_PATH_CONFIG_NAME: "/",
     state.HEALTH_CHECK_INTERVAL_CONFIG_NAME: 30,
     state.BACKENDS_PATH_CONFIG_NAME: "/",
     state.PROXY_CACHE_VALID_CONFIG_NAME: '["200 302 1h", "404 1m"]',
@@ -136,7 +135,6 @@ def test_integration_data(charm: ContentCacheBackendsConfigCharm, harness: Harne
         "protocol": "https",
         "backends_path": "/",
         "health_check_interval": "30",
-        "health_check_path": "/",
         "proxy_cache_valid": '["200 302 1h", "404 1m"]',
     }