+++ tests/unit/test_state.py
@@ -13,7 +13,6 @@ from errors import ConfigurationError
 from src.state import (
     BACKENDS_CONFIG_NAME,
     BACKENDS_PATH_CONFIG_NAME,
-    HEALTH_CHECK_PATH_CONFIG_NAME,
     HOSTNAME_CONFIG_NAME,
     PATH_CONFIG_NAME,
     PROTOCOL_CONFIG_NAME,
@@ -53,7 +52,6 @@ def test_valid_config():
     assert config.path == "/"
     assert config.backends == (IPv4Address("10.10.1.1"), IPv4Address("10.10.2.2"))
     assert config.protocol == "https"
-    assert config.health_check_path == "/"
     assert config.health_check_interval == 30
     assert config.backends_path == "/"
     assert config.proxy_cache_valid == ()
@@ -73,7 +71,6 @@ def test_hostname_with_subdomain():
     assert config.path == "/"
     assert config.backends == (IPv4Address("10.10.1.1"), IPv4Address("10.10.2.2"))
     assert config.protocol == "https"
-    assert config.health_check_path == "/"
     assert config.health_check_interval == 30
     assert config.backends_path == "/"
     assert config.proxy_cache_valid == ()
@@ -133,7 +130,6 @@ def test_longer_path():
     """
     charm = MockCharmFactory()
     charm.config[PATH_CONFIG_NAME] = "/path/to/destination/0"
-    charm.config[HEALTH_CHECK_PATH_CONFIG_NAME] = "/path/to/destination/1"
     charm.config[BACKENDS_PATH_CONFIG_NAME] = "/path/to/destination/2"
 
     config = Configuration.from_charm(charm)
@@ -141,7 +137,6 @@ def test_longer_path():
     assert config.path == "/path/to/destination/0"
     assert config.backends == (IPv4Address("10.10.1.1"), IPv4Address("10.10.2.2"))
     assert config.protocol == "https"
-    assert config.health_check_path == "/path/to/destination/1"
     assert config.health_check_interval == 30
     assert config.backends_path == "/path/to/destination/2"
     assert config.proxy_cache_valid == ()
@@ -176,24 +171,6 @@ def test_invalid_path():
     assert (
         str(err.value)
         == "Config error: ['path = /^: Value error, Path contains non-allowed character']"
-    )
-
-
-def test_invalid_health_check_path():
-    """
-    arrange: Mock charm with path with invalid character.
-    act: Create the configuration from the charm.
-    assert: Correct configurations from the mock charm.
-    """
-    charm = MockCharmFactory()
-    charm.config[HEALTH_CHECK_PATH_CONFIG_NAME] = "/path/to/`"
-
-    with pytest.raises(ConfigurationError) as err:
-        Configuration.from_charm(charm)
-
-    assert (
-        "'health_check_path = /path/to/`: Value error, Path contains non-allowed character'"
-        in str(err.value)
     )
 
 
@@ -259,7 +236,6 @@ def test_http_protocol():
     assert config.path == "/"
     assert config.backends == (IPv4Address("10.10.1.1"), IPv4Address("10.10.2.2"))
     assert config.protocol == "http"
-    assert config.health_check_path == "/"
     assert config.health_check_interval == 30
     assert config.backends_path == "/"
     assert config.proxy_cache_valid == ()
@@ -417,7 +393,6 @@ def test_valid_proxy_cache_valid():
     assert config.path == "/"
     assert config.backends == (IPv4Address("10.10.1.1"), IPv4Address("10.10.2.2"))
     assert config.protocol == "https"
-    assert config.health_check_path == "/"
     assert config.health_check_interval == 30
     assert config.backends_path == "/"
     assert config.proxy_cache_valid == ("200 302 30m", "400 1m", "500 1m")
@@ -438,7 +413,6 @@ def test_configuration_to_data():
         "path": "/",
         "backends": '["10.10.1.1", "10.10.2.2"]',
         "protocol": "https",
-        "health_check_path": "/",
         "health_check_interval": "30",
         "backends_path": "/",
         "proxy_cache_valid": "[]",